Upgrade Simon
Xử lý va chạm khi ăn Item, update framework

+++ GAME/Scene.h
@@ -9,7 +9,7 @@ public:
 	Camera * camera;
 	Scene() {};
 	virtual void LoadResources() =0;			// Tải resources trước khi tạo Scene
-	virtual void Update(float dt) =0;			// Update các thông số sau khoảng delta-time
+	virtual void Update(DWORD dt) =0;			// Update các thông số sau khoảng delta-time
 	virtual void Render() =0;					// Render các UI Components lên Scene
 	virtual void OnKeyDown(int keyCode) =0;		// Xử lí Scene khi nhấn phím
 	virtual void OnKeyUp(int keyCode) =0;		// Xử lí Scene khi thả phím
